Port the crypto module to OpenSSL 1.1
Replace all SSL variables on stack by variables on heap because their
types became opaque.

Do not define locking callbacks because they aren't necessary anymore.

Use the chacha20_poly1305 cipher implemented in libssl itself and not
in an external patch (this change isn't backward compatible because
the libssl implements RFC 7539 and the former implemetation was for
some early draft).